patch (with AI help) based on IDF

+++ esp32h21/src/aes/iv_mem.rs
@@ -12,7 +12,7 @@ impl W {}
 #[doc = "The memory that stores initialization vector\n\nYou can [`read`](crate::Reg::read) this register and get [`iv_mem::R`](R). You can [`reset`](crate::Reg::reset), [`write`](crate::Reg::write), [`write_with_zero`](crate::Reg::write_with_zero) this register using [`iv_mem::W`](W). You can also [`modify`](crate::Reg::modify) this register. See [API](https://docs.rs/svd2rust/#read--modify--write-api)."]
 pub struct IV_MEM_SPEC;
 impl crate::RegisterSpec for IV_MEM_SPEC {
-    type Ux = u8;
+    type Ux = u32;
 }
 #[doc = "`read()` method returns [`iv_mem::R`](R) reader structure"]
 impl crate::Readable for IV_MEM_SPEC {}
