+++ esp32h21/src/ecdsa/s_mem.rs
@@ -12,7 +12,7 @@ impl W {}
 #[doc = "The memory that stores s.\n\nYou can [`read`](crate::Reg::read) this register and get [`s_mem::R`](R). You can [`reset`](crate::Reg::reset), [`write`](crate::Reg::write), [`write_with_zero`](crate::Reg::write_with_zero) this register using [`s_mem::W`](W). You can also [`modify`](crate::Reg::modify) this register. See [API](https://docs.rs/svd2rust/#read--modify--write-api)."]
 pub struct S_MEM_SPEC;
 impl crate::RegisterSpec for S_MEM_SPEC {
-    type Ux = u8;
+    type Ux = u32;
 }
 #[doc = "`read()` method returns [`s_mem::R`](R) reader structure"]
 impl crate::Readable for S_MEM_SPEC {}
